+++ provenance_lib/tests/test_parse.py
@@ -18,7 +18,7 @@ from ..parse import (
     archive_not_parsed, select_parser, parse_provenance,
 )
 from ..util import UUID
-from ..archive_parser import (
+from .._archive_parser import (
     ParserV0, ParserV1, ParserV2, ParserV3, ParserV4, ParserV5,
     Config, ProvNode, ParserResults, ArtifactParser,
 )
